pyproject.toml
[tool.poetry]
name = "currency-converter"
version = "0.1.0"
description = "Currency converter from open api source https://www.exchangerate-api.com"
authors = ["Nikita <nvk.mpei@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
